requirements.txt
black==24.8.0
ipython
matplotlib>=3.2.2
numpy>=1.18.5,<1.24.0
opencv-python>=4.1.1
paho-mqtt==2.1.0
pandas>=1.1.4
protobuf<4.21.3
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
seaborn>=0.11.0
tensorboard>=2.4.1
thop
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0